Pedantic fixes to the client, where I alphabetized all of the include statements, as well as removing the new skill dialog, which we do not, nor will we use (if we do, it'd be a new one that we'd make).
WARNING!!! This, and all other previous builds have a linker error for 
the Gnome libraries version 4.3.2 on my setup. It's assumed that this is 
also the case for other users of this library as well. I'm currently 
assuming that there's a bug in the compiler itself, and will look into 
reporting this, but in the mean time, it doesn't build for these users, 
unfortunately. Sorry about this.