Fixed wrapping in some cases, removed draw method
The 'draw' method was confusingly named, and was actually for updating
the GUI after a resize. Its functionality has been merged into
'widgetResized'.

The wrapping was broken in the case where text was added to the NPC
dialog, cause it was added to an already wrapped string, causing the
wrapping to become permanent.

Sorry for all the reformatting.

Signed-off-by: Bjørn Lindeijer <bjorn@lindeijer.nl>