Fixed inconsistency in default window sizes
The default window sizes were all too small, since they were specified
in content size. On pressing "Reset Windows", the sizes would be
interpreted as such and apply alright.

The inconsistency is now removed, and the default window sizes are
always the size of the whole widget now, not just the contents.

Signed-off-by: Bjørn Lindeijer <bjorn@lindeijer.nl>