An attempt to fix the crashes related to particles
The player character is never deleted, and hence might have particles
still active when the map changes. These particles are deleted on map
change, but when the player character was moved, it was trying to mark
these deleted particles for deletion, writing to unallocated memory.

The marking for deletion by the player character now happens before the
particles are deleted.

Signed-off-by: Bjørn Lindeijer <bjorn@lindeijer.nl>