Tweaked some configuration defaults
Scrolling is now a bit less lazy and the scroll radius is 0. In my
opinion this is a better default.

Framerate limiter is now off by default, since it makes the game appear
a bit choppy. It's basically only useful for laptops anyway, and not too
hard to find.

(cherry picked from eAthena branch,
 commit c56bc78b5f5676784265dff8ed5334884e8dcc39)

Conflicts:

	src/main.cpp