Fixed up the copyright headers
Decision is to go with exact authors when some file is basically done by
one person, and in most other cases assign TMW or Aethyra team copyright
when multiple people from those teams have been involved.

Also, the first summary line should now contain the subset of
functionality that the file is part of, not the name of the whole
project.