Fixed inability to chat while talking to NPCs
These checks were meant to allow the keyboard to be used to close the
NPC dialogs, however in this form they prevent the user from being able
to chat while talking to NPCs.

Better make sure the Ok button is focused, so you can use Space to close
the NPC dialog.