Extending the internal handling of colors
The internal storage for colors was in the file color.h/color.cpp. It
mainly managed the colors in the chat.

The Color class was extended to be more generic now and it stores
gcn::Color objects instead of integers now. A lot of new colortypes are
now available, though not many of them are used for now, that will come
in the next patches.

The Color class was renamed to Palette and color.{h,cpp} to
palette.{h,cpp} to better describe its purpose.

The color config gui now lists the new colors, even changes them, but the
result is not displayed properly for now.