Modified skin loading to save and load a skin's XML path, as well as modified the skin loading method to take a default value, in case the value in the configuration file fails to load for one reason or another.
While this doesn't directly expose skinning on a per window basis to the
user at the moment, it does allow people to change what skins get loaded
with which windows now without needing to modify the code.

TODO: Determine a decent approach to allowing the user to change their
window skins in game, as well as moving all widget skin loading to the
skin class (for instance, the button skins, progressbar skins, etc.) so
that different skin configurations can use different widget skins.

Signed-off-by: Ira Rice <irarice@gmail.com>