First batch of include cleanup
Since the relative includes are getting ugly and somewhat inconvenient,
we'll switch to includes relative to the project root. We've done this
for a while already with tmwserv.

Another rule which we've always had but was never written down, is that
each source file first includes its own header. This is common practice
and ensures each header compiles without needing other stuff to be
included first.

Somebody using Code::Blocks might want to fix the project file by adding
the 'src' directory to the list of include paths.