Take the server port out of the updates dir
It doesn't make much sense to combine the update host with the server
port. Either use the update host (and path) or the server and its port.
To me it makes more sense that if different servers share the same
update host, this data should be shared by the client.

Ultimately we should store the updates by their hash, so that we can
stick everything in the same place and avoid storing any duplicates.