Replaced emote window with a new emote popup
The popup is meant to be temporary, showing the emote instantly when you
click on it. It adapts nicely to show any number of emotes.

I still need to add a better way of assigning emote shortcuts, since
currently you can only change which emote is assigned to which shortcut
by actually using it.