Fixed the empty item link crash differently
The crash was due to an assertion which shouldn't have been there, so I
removed the assertion instead. I've also made sure the unknown item has
its id initialized to 0, so that it can be used to check against instead
of the item name.

Normalization of item names was moved within the item database.