Fixed compile warnings and crash in skill dialog
The crash was probably due to roaming or uninitialized pointers in the
skill map. Please remember to clear it after deleting its contents, and
also remember that operator[] is non-const and will insert an element
when it's not already there. The proper way to see if a map contains an
element is by using find().