Introduced Net::LoginHandler::SetEmailOnRegister
This "optional action" specifies whether the server expects to get an
email address during registration. It is used now instead of having the
general handlers of eAthena and tmwserv set a pointer to an email
string on the GUI dialogs (to keep things understandable, the dependency
should preferably go one way).