Fixed crash when failing to switch to fullscreen/windowed
The setup page tries to restore the previous video mode by calling
setFullscreen again after it has failed. However, the setFullscreen
function assumed a valid mode was set (it referenced mScreen). This
would then crash.

Worked around by remembering the parameters passed to setVideoMode, and
using those in setFullscreen.

Besides fixing a potential crash, this also fixes switching between
fullscreen and windowed on Maemo 5. Probably trying to keep the color
depth the same was what made it fail (which is not necessary anyway).