Use MYPICTURES on windows, add config parameter for screenshot directory
Under Windows try to place the screenshots in the `My Pictures' directory,
and use Desktop as a fallback.

On all platforms, add a Suffix to the screenshot directory path to avoid
cluttering the Desktop/...

Usage of suffix/suffix content/path are configurable.

Create the screenshot directory if it does not exist, and fall back to saving
screenshots in the users home directory if it could not be created.

Reviewed-by: Jared Adams