Reviewed the LocalPlayer::nextTile() function. The ManaServ movement system is functional!!
Introduced LocalPlayer::getNextWalkPosition(unsigned char dir)
which takes care about the player next position while moving using keyboard.
I removed the pixel scaler thing because it couldn't handle all the noticed cases
and was rather heavy.

There is still a bug in the movement system (nothing's perfect) but it's very rare
and this is here in eAthena, too.

So, I'll give a try at taking care of it once I'll have polished all of this a bit.

Please try and give feedback!!