Disabled languages where appropriate
Languages with less than half of the strings translated were disabled
because this would make the interface very inconsistent.

Other languages were disabled because their characters are not supported
in the font or because they are written right-to-left, which is
currently not supported. This part was done by Kess Vargavind.

Task-number: 707 (TMW Mantis)
Reviewed-by: Jared Adams