Fixed the dyecmd tool.
The bug was quite nasty to catch:
The RGBA channels were inversed when writing the image, even if the code parts
taken from the Mana Client were exactly copy/pasted.
This was due to the fact that the client does an Image::load() call after recoloring
which call SDL_displaySurface() that reverse the pixel order.
I reversed then the recolored image pixel output to get back the right color
for the image writer.

Reviewed-by: Jaxad0127