Add error.xml to client repository and move error.png
There is no reason why error.xml shouldn't be in this repository, as it is used
when no sprite data is present and is a very small file anyways. Servers can
still replace these if they want to. error.png is only used in sprites, so
should be in the new sprites folder too.

Reviewed-by: 4144