Fixing segmentation fault if server is given on command line
If server and port was given on the command line, then the server type
was unknown. The command line options do work now, but only if standard
ports 6901 and 9601 are used.

This resolves http://bugs.manasource.org/view.php?id=177

TODO: Query the server about itself and choose the server type based on that.

Reviewed-by: Jaxad0127