Fixing reconnection after lost connection
An orderly disconnect of the map-server does work fine. Then in
Game::logic() an OkDialog is created, which triggers an event that
set the state to STATE_CHOOSE_SERVER.

Upon an unexpected disconnect the same OkDialog is created, but the player
does not see it. That is because the state gets changed to STATE_ERROR,
which creates another OkDialog. It changes the state to STATE_CHOOSE_SERVER
as well, but the ServerDialog does not take any input.

Reviewed-by: Jaxad0127