Fixing unusable config.xml
If config variable screenshotDirectory has national symbols in it, the
client could not read cofig.xml, and always used default settings.

In order to get ready for a client release, now the config.xml does not
save the screenshot folder. The --screenshot-dir commandline option
is working as before, the value is just not permanent anymore. This
makes actually sense, since the options for update folder and config
folder are not permanent either.

This resolves http://bugs.manasource.org/view.php?id=112

Reviewed-by: Jaxad0127