Implementing show IP for game masters
As an upcoming feature the TMW-Athena server sends IP addresses or
IP hashes to game masters. The current client freezes if it receives
such a packet, therefor the game masters need to use a new client before
the server can use it. Normal players are not affected, because they
do not get this packet.

Showing the IP is optional and can be enable with the chat command
"/showip 1". The IP is then shown behind the players name.

Reviewed-by: Bertram