Updating F1-help
Based on Bertram's start some more updates to the F1-help.

The team page was removed, because the last update was done in
March 2006 and no one volunteered to write a new one.

TODO: Improvements can be done during the beta test. A team list
can be added if someone wants to do it.

This resolves http://bugs.manasource.org/view.php?id=210