Adding theme colors for party tabs
Following Rotonen's advise to "make anything and everything you
can think of recolorable", a theme can now define colors for
PARTY_CHAT_TAB and PARTY_SOCIAL_TAB. This was badly needed, because
the requirements for a color of a chat tab and of a player name
are quite different.

TODO: When guild support gets enabled, then have a look at the
colors for guilds.

Reviewed-by: 4144