Basically merged the two movement algorithms into one.
This was made in favour of the manaserv way of doing things.

I also added a way to keep the original server speed value
so the pixel value can be recomputed at each map change, as this
was necessary since the speed is given before the first map
is loaded.

The code is much more simpler now about movement handling,
and we can already see improvements on other characters
movements in The Mana World with this.

Everything can't be perfect the first time; here are bugs
identified so far:
- Monsters direction isn't updated on TmwAthena for obscure reasons.
- Remote players walking animation is sometimes reset on each steps.
- When changing map, the local player sometimes walks randomly
until the player reacts.

Stay tuned!