Made the client load the skills.xml again.
This, instead of the protocol specific filenames that are
mana-skills.xml and ea-skills.xml.

This is backward compatible as the old skills.xml file was once
used before the 0.0.29.1 release which is the last alive from
the 0.0.x serie.

The skills.xml will have to added to world data
to make this working, though. (Done in another patch.)