Fix attack speed on both servers.
Currently, we actually receive the local player attack speed
in milliseconds for tA.
Yet, we don't support receiving the attack speed for other beings,
so the old behaviour will remain for them until someone adds that.

As for Manaserv, we're only using the default speed atm.

Most urgent part of mana-mantis: #343.

Reviewed-by: Thorbjorn.