Added effects support in monsters attacks.
This is provided through the use of the 'effect-id' property
of the <attack> tag.
Note that the 'particle-effect' support has been removed
as it was redundant, but can be kept for backward compatibility.

Now, on both Manaserv and tA, only the attack id 1 is supported,
i.e:
<monster>
  ...
  <attack id="1" effect-id="11" action="attack" />
</monster>

More to come with re-enabling of the attack type support.

I had to fix a crash in the Being::takeDamage() function
when there were no attacker (in Manaserv), and add rotation
support to the effectManager to achieve this.

Reviwed-by: Jaxad0127.