Made the being death sequence better handled.
Now the death animation is set to the minimum of 1.5 seconds
just as in the older versions, but permit longer sequences
as well.

I didn't put a maximum hard cap on it as it may be a burden
for the maintainers whatever the hard cap would be,
and as it isn't vital to set one anyway.

Resolves: Mana-Mantis #364.

Reviewed-by: bjorn, Jaxad0127.