Added a close button to the npc dialog window.
The close button will only display when the window is waiting
for the server, so that crashed windows can be closed without
restarting the client.
Pushing the QUIT key (Escape by default) will also close
the crashed windows in that case.

I also fixed a memleak with the "next" button.

Resolves: Mana-Mantis #72, 93, 389.

Reviewed-by: Ablu.