Fixed the chat bubbles position.
This issue was much nastier than it first sounded.

The issue was happening when using a long weapon (like the scythe)
with the SDL backend.
The algorithm computing the sprite compound frames into one
would then update the sprite height to a high value
(127 in my case), leading to put the bubble too far away
from the player.

The algorithm not being in cause about the needed height,
I noticed that the simplest way to set a good Y position
of the text bubbles was to simply set a maximum.
A config option can be added later.

I also unified the way the position is computed in the being class.

+ Function description sentence removal requested by bjorn.

Resolves: Mana-Mantis #447.

Reviewed-by: bjorn.