Found a better way to fix the movement glitches on both servers.
The patch also takes care of not spamming the different servers,
when the servers are setting the being speed correctly.

The most problems were coming from the keyboard movement functions
handling 1 tile paths. To void the issues seen in #405, #439,
and #440, I simply prevented to set a new path before reaching
the destination of the former one, when using the keyboard.

The mouse path system remains unchanged.

I also made some functions private (or here protected)
to show they shouldn't be called by something else than
the localplayer object.

And I removed the nextTile() function, since it was obsolete,
unused, and replaced by the nextTile(direction) function.

That patch was tested on both servers with mouse/keyboard
mixed use.

Resolves: Mana-Mantis #405, #439, #440.

Reviewed-by: bjorn