Fixed the click area for NPCs and other players
By taking into account that the sprite is drawn half a tile lower than
the position of the being (that is messy on its own, but hard to fix
while supporting both pixel and tile based movement).

It also makes sure that the click area is at least 32x32 pixels, no
matter how small the being. This fixes the lack of a clickable area
for NPCs without a sprite (like some chests or signs) and makes
small monsters like maggots easier to hit.

Reviewed-by: Yohann Ferreira