Fixed diagonal walking to be much more smooth.
The character used to stick to both corners of an obstacle it
encountered when walking diagonally.
Now it simply go back to the straight moving mode when
encountering an obstacle in one of the two direction used.

This also simplifies the function logic.

The character direction bug left is still there,
and will be dealt in a separate patch with.

Reviewed-by: Erik Schilling