Improved the wallpaper choosing algorithm
Now it searches for the smallest wallpaper that is at least as large as
the screen, since downscaling is preferred to upscaling. When not found,
it will go with the largest available wallpaper.

Previously, the algorithm preferred upscaling, but didn't bother
searching for the largest available one, but would pick a random image.
It would also revert to the default Mana wallpaper when no image was
small enough to fit on the screen.

Reviewed-by: Erik Schilling
Reviewed-by: Yohann Ferreira