Added notification sound on receiving whisper
One of the sound channels is reserved for notification sounds, of which the
volume can be configured separately. Currently, the only notification sound
that is played is for receiving whispers. That can be extended later.

The newmessage.ogg sound used currently is the one for receiving a message
with the Psi instant messenger.

Parts of this patch are based on the new message notification in ManaPlus.

Reviewed-by: Erik Schilling