Fixed category of mana.desktop file
It is not allowed to provide multiple different categories since it would spam
the menu too much. openSUSE prevents packaging if a .desktop file does not
follow this rule. So I removed the Application category and added a subcategory
of the Game category.

See http://standards.freedesktop.org/menu-spec/latest/apa.html for more
information.

Reviewed-by: Bertram, jurkan.