Removed some 'const' keywords from value returns
It makes no sense to mark a return value as const when it is returned by
value.  This only makes sense if the return value is passed by reference, in
order to prevent the receiver from modifying the value.

Also made some other small adjustments. A std::string does not need to be
initialized to "" explicitly and getSafeUtf8String could take its parameter by
reference.

Reviewed-by: Yohann Ferreira