Potentially fixed problems with user agent in Windows builds
The Windows releases have not been properly identifying themselves, and
it seems to be due to a compiler issue, as discovered by Stefan Dombrowski
(@cody). He did a similar change but didn't push it anywhere so I've redone
it.

The line that is now split up was very preprocessor-heavy, which may have
been part of the problem, but we'll probably never know what it was
exactly.