Fixed problems with using the last emote
The tmwa/BeingHandler was adjusting the effect id rather than the emote
id, causing it to not find the last emote (and the rest of the emotes
only worked correctly because the effect ids were consecutive in the same
order as the emote ids).

Furthermore, the EmoteShortcutContainer refused to draw the icon for the
last emote due to an off-by-one error in dealing with the 1-adjusted
emote ids used by the EmoteShortcut class.

Also cleaned up some old remains of a player ignore strategy that used
to use the two balloon emotes (this had been their original purpose).

Reviewed-by: Erik Schilling