Added emote support for manaserv-based server - client part
Due to the fact that the server won't handle id equal to zero,
I added a new bound check when loading the emotes
in the emotepopup window.

Note that the emote shortcut bar is still not functioning due
to the offset problem remaining for tA servers,
but I'd rather see that fixed in another commit.