Removed manual calling of glFlush and made glFinish optional
Actually neither glFlush nor glFinish need to be called manually [1].
However, the rendering pipeline is then free to queue up future frames
which can cause a lag which is especially noticable with mouse movement.

To avoid this lag, we call glFinish after SDL_GL_SwapBuffers. This makes
sure processing of the next frame does not start before the current
frame is being displayed.

[1] http://www.opengl.org/wiki/Common_Mistakes#glFinish_and_glFlush