Avoid copying surfaces unnecessarily for upload to texture
When uploading an SDL surface to an OpenGL texture, it was always making
a copy that had the desired size and pixel format. Now this copy is no
longer being made when the existing surface already has the target size
and pixel format.

In fact most images are already in 32-bit RGBA format after loading them.

Reviewed-by: Erik Schilling