Added support for charcreation.xml settings
This file was introduced by ManaPlus as a way of configuring the
character creation process. It defines the number of hair styles
and colors, how stat points should be divided and what the starting
equipment of the player is.

The minimum and maximum hair color/style IDs are not supported at
the moment.

This is mostly a backport of ManaPlus commits 10cf52b5 and dcc18eba,
with some style changes.

Mantis-issue: 501
Reviewed-by: Ablu